Refactoring to include sub-commands (aws cli style)
+ introduce milestone top-command (@see https://github.com/yargs/yargs/issues/225)
+ use yargs command module pattern (@see https://github.com/yargs/yargs/blob/master/docs/advanced.md#providing-a-command-module)
