+++ lib/auth.js
@@ -78,7 +78,7 @@ function hydrateMilestones(github) {
                     repo.milestones = response.data
                         .map(m => {
                             return {
-                                id: m.id, 
+                                id: m.id,
                                 number: m.number,
                                 title: m.title.trim(),
                                 state: m.state
@@ -87,7 +87,7 @@ function hydrateMilestones(github) {
                         .sort(sortByNumber)
                 }
                 progress.tick();
-            }));       
+            }));
         }
         log.success(`${count} total milestones found.`);
     };
