+++ lib/commands/milestone/change-status.js
@@ -1,6 +1,6 @@
 const ProgressBar = require('progress');
-const auth = require('../auth');
-const log = require('../log');
+const auth = require('../../auth');
+const log = require('../../log');
 
 module.exports = state => {
     return async function (argv) {
@@ -10,7 +10,7 @@ module.exports = state => {
             const milestones = await filterMilestones(repos, argv.milestone);
             if (!milestones.length) return;
 
-            log.info(`Changing status to ${state}...`);
+            log.info(`Changing status to ${state}...`);Ò
             const progress = new ProgressBar(':current/:total [:bar]', { total: milestones.length });
 
             for (let milestone of milestones) {
