+++ lib/commands/milestone/close.js
@@ -1,0 +1,7 @@
+const changeStatus = require('./change-status');
+
+module.exports.command = 'close [milestone]';
+
+module.exports.describe = 'Close milestone';
+
+module.exports.handler = changeStatus('closed');
