+++ lib/commands/milestone/index.js
@@ -1,0 +1,7 @@
+module.exports = yargs => {
+    return yargs
+        .command(require('./list'))
+        .command(require('./close'))
+        .command(require('./open'))
+        .command(require('./update'));
+};
