+++ lib/commands/milestone/list.js
@@ -1,9 +1,22 @@
 const ProgressBar = require('progress');
 const colors = require('colors');
-const auth = require('../auth');
-const log = require('../log');
+const auth = require('../../auth');
+const log = require('../../log');
 
-module.exports = async function(argv) {
+module.exports.command = 'list';
+
+module.exports.describe = 'List milestones in all your repositories';
+
+module.exports.builder = {
+    state: {
+        alias: 's',
+        type: 'state',
+        choices: ['all', 'open', 'closed'],
+        describe: 'state of the milestone'
+    }
+};
+
+module.exports.handler = async function(argv) {
     const { github, repos, hydrateMilestones, filterMilestones } = await auth.init();
     await hydrateMilestones(repos, argv.state);
 
@@ -20,4 +33,4 @@ module.exports = async function(argv) {
             console.log(`• ${m.number}: ${color(m.title)} (${color(m.state)})`);
         }
     }
-}
+};
