+++ lib/commands/milestone/open.js
@@ -1,0 +1,7 @@
+const changeStatus = require('./change-status');
+
+module.exports.command = 'open [milestone]';
+
+module.exports.describe = 'Re-open milestone';
+
+module.exports.handler = changeStatus('open');
