+++ lib/commands/milestone/update.js
@@ -1,8 +1,30 @@
 const ProgressBar = require('progress');
-const auth = require('../auth');
-const log = require('../log');
+const auth = require('../../auth');
+const log = require('../../log');
 
-module.exports = async function(argv) {
+module.exports.command = 'update [milestone]';
+
+module.exports.describe = 'Update milestone';
+
+module.exports.builder = {
+    force: {
+        alias: 'f',
+        type: 'boolean',
+        describe: 'force update on closed milestones as well'
+    },
+    title: {
+        alias: 't',
+        type: 'string',
+        describe: 'title to change to'
+    },
+    due: {
+        alias: 'd',
+        type: 'string',
+        describe: 'milestone\'s due date. format: YYYY-MM-DD'
+    }
+};
+
+module.exports.handler = async function(argv) {
     const { github, repos, hydrateMilestones, filterMilestones } = await auth.init();
     await hydrateMilestones(repos, argv.force ? 'all' : 'open');
     const milestones = await filterMilestones(repos, argv.milestone);
@@ -25,4 +47,4 @@ module.exports = async function(argv) {
     }
 
     log.success(`Done.`);
-}
+};
