+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "ghm-cli",
   "version": "1.0.3",
-  "description": "Github multi-repo milestone management CLI",
+  "description": "Github multi-repo management CLI",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
